REFACTOR: Move plugins to openbci dir
