+++ openbci/__init__.py
@@ -1,8 +1,9 @@
 
 from .cyton import OpenBCICyton
 from .ganglion import OpenBCIGanglion
-from .wifi import OpenBCIWiFi
+from .plugins import *
 from .utils import *
+from .wifi import OpenBCIWiFi
 
 
 __version__ = "1.0.0"
