+++ openbci/plugins/__init__.py
@@ -1,0 +1,9 @@
+
+from .csv_collect import *
+from .noise_test import *
+from .streamer_lsl import *
+from .streamer_osc import *
+from .streamer_tcp_server import *
+from .udp_server import *
+
+__version__ = "1.0.0"
