+++ openbci/plugins/sample_rate.py
