+++ setup.py
@@ -3,8 +3,8 @@ from setuptools import setup, find_packages
 setup(name = 'OpenBCI_Python',
       version = '1.0.0',
       description = 'A lib for controlling OpenBCI Devices',
-      author='Jeremy Frey',
-      author_email='frederic.simard.1@outlook.com',
+      author='AJ Keller',
+      author_email='pushtheworldllc@gmail.com',
       license='MIT',
       packages=find_packages(),
       install_requires=['numpy'],
